Rename KShortestPathsExecutor files to EnumeratePathsExecutor
