+++ tests/CMakeLists.txt
@@ -128,7 +128,7 @@ set(ARANGODB_TESTS_SOURCES
   Aql/IndexNodeTest.cpp
   Aql/InputRangeTest.cpp
   Aql/InsertExecutorTest.cpp
-  Aql/KShortestPathsExecutorTest.cpp
+  Aql/EnumeratePathsExecutorTest.cpp
   Aql/EnumeratePathsNodeTest.cpp
   Aql/LimitExecutorTest.cpp
   Aql/MockTypedNode.cpp
